LRCI-378 Extract single suite specific methods into the single suite class
